% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/addMetrics.R
\name{addMetrics}
\alias{addMetrics}
\alias{addMetrics,SummarizedExperiment-method}
\title{Add Multiple Metrics (e.g., CRT) to a \code{\link{SummarizedExperiment}} object}
\usage{
addMetrics(x, ...)

\S4method{addMetrics}{SummarizedExperiment}(
  x,
  assay.type = "counts",
  metrics = c("CRT"),
  thresholds = list(CRT = 2),
  split.by = NULL,
  ...
)
}
\arguments{
\item{x}{a \code{\link{SummarizedExperiment}} object.}

\item{...}{Additional arguments to pass to individual metric functions.}

\item{assay.type}{\code{Character scalar}. Specifies the name of the assay
used in calculation. (Default: \code{"counts"})}

\item{metrics}{\code{Character vector}. Specifies the metrics to be
calculated. For example, \code{CRT}. (Default: \code{c("CRT")})}

\item{thresholds}{\code{Named list}. A list of thresholds for each metric.
For example, \code{list(CRT = 2)} specifies the threshold for CRT
calculation. (Default: \code{list(CRT = 2)})}

\item{split.by}{\code{Character scalar}. Specifies how to split the data
for per-group or per-subject calculation. For example, \code{"subject"}.
If \code{NULL}, no splitting will occur. (Default: \code{NULL})}
}
\value{
A \code{\link{SummarizedExperiment}} object with added metadata
containing the calculated metrics.
}
\description{
Add Multiple Metrics (e.g., CRT) to a \code{\link{SummarizedExperiment}} object
}
