% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addOrderSequence.R
\docType{methods}
\name{addOrderSequence}
\alias{addOrderSequence}
\alias{addOrderSequence-SummarizedExperiment}
\alias{addOrderSequence,SummarizedExperiment-method}
\title{Rank the sample information in a \code{SummarizedExperiment} Object}
\usage{
addOrderSequence(data, field, rank_field_name, ...)
}
\arguments{
\item{data}{: \linkS4class{SummarizedExperiment} object containing time
series.}

\item{field}{: A character value indicating the name of the matrix
column that is aimed to be ranked.}

\item{rank_field_name}{: A character value to name the newly ranked \code{field}}

\item{...}{Allow new parameters to be defined for this function.}
}
\value{
\linkS4class{SummarizedExperiment} object with ranked field
added
}
\description{
The information stored in \code{colData} in
\linkS4class{SummarizedExperiment} object is ranked by \code{addOrderSequence}
function and newly ranked field is added back to
\linkS4class{SummarizedExperiment} object.
}
\examples{
data(GlobalPatterns, package="mia")
se <- GlobalPatterns

se1 <- addOrderSequence(se,field = "SampleType",
            rank_field_name = "SampleType_rank",
            na.last = TRUE, ties.method = "first")

data(hitchip1006)
hitchip_data <- hitchip1006
hitchip_data <- addOrderSequence(hitchip_data, field = "time",
            rank_field_name = "time_rank")

#example 2: new matrix and its rank added

se$newfield <- matrix(1:26, nrow = 26, ncol = 1)
se2 <- addOrderSequence(se, field = "newfield" ,
                rank_field_name = "newfield_rank" , na.last = TRUE ,
                ties.method = "first")

}
