% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getShortTermChange.R
\name{addShortTermChange}
\alias{addShortTermChange}
\alias{getShortTermChange}
\alias{getShortTermChange,SummarizedExperiment-method}
\alias{addShortTermChange,SummarizedExperiment-method}
\title{Short term Changes in Abundance}
\usage{
addShortTermChange(x, ...)

getShortTermChange(x, ...)

\S4method{getShortTermChange}{SummarizedExperiment}(x, assay.type = "counts", ...)

\S4method{addShortTermChange}{SummarizedExperiment}(x, assay.type = "counts", name = "short_term_change", ...)
}
\arguments{
\item{x}{a \code{\link{SummarizedExperiment}} object.}

\item{...}{additional arguments.}

\item{assay.type}{\code{Character scalar}. Specifies the name of assay
used in calculation. (Default: \code{"counts"})}

\item{name}{\code{Character scalar}. Specifies a name for storing
short term results. (Default: \code{"short_term_change"})}
}
\value{
\code{getShortTermChange} returns \code{DataFrame} object containing
the short term change in abundance over time for a microbe.
\code{addShortTermChange}, on the other hand, returns a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object with these results in its \code{metadata}.
}
\description{
Calculates short term changes in abundance of taxa
using temporal Abundance data.
}
\details{
This approach is used by Wisnoski NI and colleagues
\url{https://github.com/nwisnoski/ul-seedbank}. Their approach is based on
the following calculation log(present abundance/past abundance).
Also a compositional version using relative abundance similar to
Brian WJi, Sheth R et al
\url{https://www.nature.com/articles/s41564-020-0685-1} can be used.
This approach is useful for identifying short term growth behaviors of taxa.
}
\examples{

# Load time series data
data(minimalgut)
tse <- minimalgut

short_time_labels <- c("74.5h", "173h", "438h", "434h", "390h")

# Subset samples by Time_label and StudyIdentifier
tse <- tse[, !(tse$Time_label \%in\% short_time_labels)]
tse <- tse[, (tse$StudyIdentifier == "Bioreactor A")]

# Get short term change
# Case of rarefying counts
tse <- transformAssay(tse, method = "relabundance")
getShortTermChange(tse, assay.type = "relabundance", time.col = "Time.hr")

# Case of transforming counts
tse <- rarefyAssay(tse, assay.type = "counts")
getShortTermChange(tse, assay.type = "subsampled", time.col = "Time.hr")
}
