% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStepwiseDivergence.R
\name{addStepwiseDivergence}
\alias{addStepwiseDivergence}
\alias{getStepwiseDivergence}
\alias{getStepwiseDivergence,ANY-method}
\alias{addStepwiseDivergence,SummarizedExperiment-method}
\title{Beta diversity between consecutive time steps}
\usage{
getStepwiseDivergence(x, ...)

\S4method{getStepwiseDivergence}{ANY}(
  x,
  time.col,
  assay.type = "counts",
  time.interval = 1L,
  group = NULL,
  method = "bray",
  name = "divergence",
  name.time = "time_diff",
  ...
)

addStepwiseDivergence(x, ...)

\S4method{addStepwiseDivergence}{SummarizedExperiment}(x, name = "divergence", name.time = "time_diff", ...)
}
\arguments{
\item{x}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{...}{optional arguments passed into
\code{\link[mia:addDivergence]{mia::addDivergence()}}.}

\item{time.col}{\code{Character scalar}. Specifies the name of the
time series field in \code{colData}.}

\item{assay.type}{\code{Character scalar}. Specifies which assay values are
used in the dissimilarity estimation. (Default: \code{"counts"})}

\item{time.interval}{\code{Integer scalar}. Indicates the increment between
time steps. If you need to take every second, every third, or so, time step
only, then increase this accordingly. (Default: \code{1L})}

\item{group}{\code{Character scalar}. Specifies the grouping
factor (name of a \code{colData} field). If given, the divergence is calculated
per group.  e.g. subject, chamber, group etc. (Default: \code{NULL})}

\item{method}{\code{Character scalar}. Used to calculate the distance.
Method is passed to the function that is specified by \code{dis.fun}.
(Default: \code{"bray"})}

\item{name}{\code{Character scalar}. Shows beta diversity between
samples. (Default: \code{"time_divergence"})}

\item{name.time}{\code{Character scalar}. Field name for adding the
time difference between samples used to calculate beta diversity.
(Default: \code{"time_difference"})}
}
\value{
a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
containing the sample dissimilarity and corresponding time difference between
samples (across n time steps), within each level of the grouping factor.
}
\description{
Calculates sample dissimilarity between consecutive time points (t, t+i),
within a group (subject, reaction chamber, or similar). The corresponding
time difference is returned as well. The method operates on
\code{SummarizedExperiment} objects, and the results are stored in \code{colData}.
}
\examples{
library(miaTime)

data(hitchip1006)
tse <- transformAssay(hitchip1006, method = "relabundance")

# Calculate divergence
tse <- addStepwiseDivergence(
    tse, group = "subject",
    time_interval = 1,
    time.col = "time",
    assay.type="relabundance"
    )

}
