% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStepwiseDivergence.R
\name{getStepwiseDivergence}
\alias{getStepwiseDivergence}
\alias{getTimeDivergence}
\alias{getTimeDivergence,ANY-method}
\title{Beta diversity between consecutive time steps}
\usage{
getStepwiseDivergence(
  x,
  group = NULL,
  time_field,
  time_interval = 1,
  name_divergence = "time_divergence",
  name_timedifference = "time_difference",
  assay_name = "counts",
  FUN = vegan::vegdist,
  method = "bray",
  ...
)

getTimeDivergence(x, ...)

\S4method{getTimeDivergence}{ANY}(x, ...)
}
\arguments{
\item{x}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{group}{optional; a single character value for specifying the grouping
factor (name of a \code{colData} field). If given, the divergence is calculated
per group.  e.g. subject, chamber, group etc.).}

\item{time_field}{a single character value, specifying the name of the
time series field in \code{colData}.}

\item{time_interval}{integer value indicating the increment between time
steps (default: 1). If you need to take every second, every third, or so, time step only, then
increase this accordingly.}

\item{name_divergence}{a column vector showing beta diversity between samples
(default: \code{name_divergence = "time_divergence"})}

\item{name_timedifference}{field name for adding the time difference between
samples used to calculate beta diversity
(default: \code{name_timedifference = "time_difference"})}

\item{assay_name}{character indicating which assay values are used in
the dissimilarity estimation (default: \code{assay_name = "counts"}).
It could also be used for \code{reducedDims} entries. Yet, the ones present in
\code{assays} are the priority.}

\item{FUN}{a \code{function} for dissimilarity calculation. The function must
expect the input matrix as its first argument. With rows as samples
and columns as features. By default, \code{FUN} is
\code{vegan::vegdist}.}

\item{method}{a method that is used to calculate the distance. Method is
passed to the function that is specified by \code{FUN}. By default,
\code{method} is \code{"bray"}.}

\item{...}{Arguments to be passed}
}
\value{
a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
or
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
containing the sample dissimilarity and corresponding time difference between
samples (across n time steps), within each level of the grouping factor.
}
\description{
Calculates sample dissimilarity between consecutive time points (t, t+i),
within a group (subject, reaction chamber, or similar). The corresponding
time difference is returned as well. The method operates on
\code{SummarizedExperiment} objects, and the results are stored in \code{colData}.
}
\examples{
#library(miaTime)
library(TreeSummarizedExperiment)

data(hitchip1006)
tse <- mia::transformSamples(hitchip1006, method = "relabundance")

# Subset to speed up example
tse <- tse[, colData(tse)$subject \%in\% c("900", "934", "843", "875")]

# Using vegdist for divergence calculation, one can pass
# the dissimilarity method from the vegan::vegdist options
# via the "method" argument
tse2 <- getStepwiseDivergence(tse, group = "subject",
                              time_interval = 1,
                              time_field = "time",
                              assay_name="relabundance",
                              FUN = vegan::vegdist,
                              method="bray")

}
