% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTsMeasure.R
\name{getTsMeasure}
\alias{getTsMeasure}
\alias{getTsMeasure,TreeSummarizedExperiment-method}
\title{Get Time Series Measures}
\usage{
getTsMeasure(x, ...)

\S4method{getTsMeasure}{TreeSummarizedExperiment}(x, time.field = "Time.hr", measures = "acf", ...)
}
\arguments{
\item{x}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{...}{Additional arguments passed to specific time series functions.}

\item{time.field}{\code{Character scalar}. Specifies the name of the
time series field in \code{colData}. (Default: \code{"Time.hr"})}

\item{measures}{\code{Character vector}. Specifies the time series measures
to compute. Supported measures include "acf", "pacf","Box.test", and "arima".
(Default: \code{ ""acf""})}
}
\value{
A list containing the results of the requested time series measures.
}
\description{
Calculate various time series measures (e.g., ACF, PACF)
from a specified column in the \code{colData} of a \code{TreeSE} object.
}
\examples{

data(minimalgut)
tse <- minimalgut
result <- getTsMeasure(tse, time.field = "Time.hr", measure = "acf")

}
