% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hitchip1006}
\alias{hitchip1006}
\title{HITChip Atlas with 1006 Western Adults}
\format{
The data set in
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
format.
}
\usage{
data(hitchip1006)
}
\value{
Loads the data set in R.
}
\description{
This data set contains genus-level microbiota profiling with
HITChip for 1006 western adults with no reported health complications,
reported in Lahti et al. (2014).
}
\details{
The data is also available for download from the Data Dryad
\url{http://doi.org/10.5061/dryad.pk75d} and was initially released
as a phyloseq object under the name atlas1006 in \pkg{microbiome} R package.
The \pkg{miaTime} package provides this as
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object.
Some of the subjects also have short time series.
}
\references{
Lahti et al. Tipping elements of the human intestinal ecosystem.
Nature Communications 5:4344, 2014. \url{https://doi.org/10.1038/ncomms5344}.
}
\author{
Leo Lahti
}
\keyword{data}
