% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mdbMM16}
\alias{mdbMM16}
\title{Human Gut Minimal Microbiome Profiling Data}
\format{
The data set in
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
format.
}
\usage{
data(mdbMM16)
}
\value{
Loads the data set in R.
}
\description{
This time-series data set contains absolute temporal abundances
for 16 human gut species that were assembled in an in vitro gut system. These
were subjected to a variety of disturbances over a period of 20 days. The
sample data includes measurements for Acetate, Butyrate, Propionate, and
optical density.
}
\details{
The data is available also on
\url{https://github.com/microsud/syncomR}
The data contains 183 samples from 3 in vitro gut system, 61 per bioreactor
collected three times daily.
The current implementation in \pkg{miaTime} provides this as
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object.
}
\references{
Shetty, S.A., Kostopoulos, I., Geerlings, S.Y. et al. Dynamic metabolic
interactions and trophic roles of human gut microbes identified using a
minimal microbiome exhibiting ecological properties.
ISME, (Accepted 2022).
}
\author{
Sudarshan A. Shetty
}
\keyword{data}
