% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortTermChange.R
\name{shortTermChange}
\alias{shortTermChange}
\alias{shortTermChange,SummarizedExperiment-method}
\title{Short term Changes in Abundance}
\usage{
shortTermChange(
  x,
  assay.type = "counts",
  rarefy = FALSE,
  compositional = FALSE,
  depth = min(assay(x, assay.type)),
  ...
)

\S4method{shortTermChange}{SummarizedExperiment}(
  x,
  assay.type = "counts",
  rarefy = FALSE,
  compositional = FALSE,
  depth = min(assay(x, assay.type)),
  plot = FALSE,
  ...
)
}
\arguments{
\item{x}{a \code{\link{SummarizedExperiment}} object.}

\item{assay.type}{\code{Character scalar}. Specifies the name of assay
used in calculation. (Default: \code{"counts"})}

\item{rarefy}{\code{Logical scalar}. Whether to rarefy counts.
(Default: \code{FALSE})}

\item{compositional}{\code{Logical scalar}. Whether to transform counts.
(Default: \code{FALSE})}

\item{depth}{\code{Integer scalar}. Specifies the depth used in rarefying.
(Default: \code{min(assay(x, assay.type)}))}

\item{...}{additional arguments.}
}
\value{
\code{dataframe} with \code{short term change} calculations.
}
\description{
Calculates short term changes in abundance of taxa
using temporal Abundance data.
}
\details{
This approach is used by Wisnoski NI and colleagues
\url{https://github.com/nwisnoski/ul-seedbank}. Their approach is based on
the following calculation log(present abundance/past abundance).
Also a compositional version using relative abundance similar to
Brian WJi, Sheth R et al
\url{https://www.nature.com/articles/s41564-020-0685-1} can be used.
This approach is useful for identifying short term growth behaviors of taxa.
}
