% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temporalMicrobiome20}
\alias{temporalMicrobiome20}
\title{Gut Microbiome Profiling 20 Belgian Women}
\format{
The data set in \code{\link{TreeSummarizedExperiment-class}}
format.
}
\usage{
data(temporalMicrobiome20)
}
\value{
Loads the data set in R.
}
\description{
This time-series data set contains absolute temporal variation
for most major gut genera, as well as diversity indicators for both
relative and quantitative abundance profiles of healthy women living in an
industrialized country.
}
\details{
The data is available also on
\url{https://www.nature.com/articles/s41467-021-27098-7#Sec43}
The data contains 713 fecal samples from 20 Belgian women collected
over six weeks.
The current implementation in \pkg{miaTime} provides this as
\code{\link{TreeSummarizedExperiment-class}} object.
}
\references{
Vandeputte, D., De Commer, L., Tito, R.Y. et al. Temporal variability
in quantitative human gut microbiome profiles and implications for clinical
research.
Nat Commun 12, 6740 (2021). https://doi.org/10.1038/s41467-021-27098-7
}
\author{
Doris Vandeputte
}
\keyword{data}
