% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeNumeric.R
\name{timeNumeric}
\alias{timeNumeric}
\title{Time field provided in character as "\%H:\%M:\%S" or in Period class ("H M S")
can be converted to numeric values depending on the units.}
\usage{
timeNumeric(time_field, units = "mins")
}
\arguments{
\item{time_field}{: vector containing the time points}

\item{units}{: The unit of time points selected in \code{time_field}.
It can be hours, minutes or seconds.(default: \code{unit = "mins"})}
}
\description{
Time field provided in character as "\%H:\%M:\%S" or in Period class ("H M S")
can be converted to numeric values depending on the units.
}
\examples{
data(hitchip1006)
hitchipTime <- timeToDate(time_field = colData(hitchip1006)[,"time"],
                                                           unit = "mins")

periodToNumeric <- timeNumeric(hitchipTime, unit = "mins")

}
