% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timePerSubject.R
\name{timePerSubject}
\alias{timePerSubject}
\title{\linkS4class{SummarizedExperiment} class object stores time information about
each sample in \code{colData} field. Time information can be categorized by
per subject(individual) and then used with the selected time operation.
The resulting vector is added as a new column of the\code{colData} field.}
\usage{
timePerSubject(se, time_field, subject_field, operation, new_field)
}
\arguments{
\item{se}{: \linkS4class{SummarizedExperiment} class object}

\item{time_field}{: column vector containing the time information}

\item{subject_field}{: column vector containing the subject(individual)
information}

\item{operation}{: a function applied to \code{time_field}}

\item{new_field}{: character class object refers to the name of a
newly added field}
}
\value{
a list object containing each subject individually with the
related data provided in the \code{colData} field
}
\description{
\linkS4class{SummarizedExperiment} class object stores time information about
each sample in \code{colData} field. Time information can be categorized by
per subject(individual) and then used with the selected time operation.
The resulting vector is added as a new column of the\code{colData} field.
}
\examples{
library(miaTime)
data(hitchip1006)
se <- hitchip1006

hitchipTime <- timePerSubject(se, time_field = "time",
    subject_field = "subject", operation ="rev", new_field = "reversed")

}
