% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimePointsToDate.R
\name{TimePointsToDate}
\alias{TimePointsToDate}
\title{Numerically provided time points are converted to a POSIXct object in the
format "\%Y-\%m-\%d \%H:\%M:\%OS".}
\usage{
TimePointsToDate(data, t, origin = "1970-01-01")
}
\arguments{
\item{data}{: dataset with time points in it}

\item{t}{: character time points field in the given dataset}

\item{origin}{: character indicating the start of the time points
(default: \code{origin = "1970-01-01"})}
}
\value{
a \linkS4class{TreeSummarizedExperiment} object containing added
date object, rank of the time points and time difference between the time
points in seconds.
}
\description{
Numerically provided time points are converted to a POSIXct object in the
format "\%Y-\%m-\%d \%H:\%M:\%OS".
}
\examples{
data(hitchip1006)
force(hitchip1006)
HitchipDate <- TimePointsToDate(hitchip1006, "time", origin = "1970-01-01")

}
