% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBaselineDivergence.R
\name{addBaselineDivergence}
\alias{addBaselineDivergence}
\alias{addBaselineDivergence,SummarizedExperiment-method}
\title{Beta diversity between the baseline and later time steps}
\usage{
addBaselineDivergence(x, ...)

\S4method{addBaselineDivergence}{SummarizedExperiment}(
  x,
  time_field,
  assay.type = "counts",
  group = NULL,
  name_divergence = "divergence",
  name_timedifference = "time_diff",
  method = "bray",
  ...
)
}
\arguments{
\item{x}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{...}{Additional parameters. See dedicated function.}

\item{baseline_sample}{\code{Character vector}. Specifies the baseline
sample(s) to be used. If the \code{group} argument is given, this must be a
named \code{vector}; one element per group.}
}
\value{
a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
or
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
containing the sample dissimilarity and corresponding time difference between
samples (across n time steps), within each level of the grouping factor.
}
\description{
Calculates sample dissimilarity between the given baseline and other
time points, optionally within a group (subject, reaction chamber, or
similar). The corresponding time difference is returned as well.
The method operates on \code{SummarizedExperiment} objects, and the results
are stored in \code{colData}.
}
\details{
The group argument allows calculating divergence per group. Otherwise, this
is done across all samples at once.

The baseline sample/s always need to belong to the data object i.e. they
can be merged into it before
applying this function. The reason is that they need to have comparable
sample data, at least some time point
information for calculating time differences w.r.t. baseline sample.

The baseline time point is by default defined as the smallest time point
(per group). Alternatively,
the user can provide the baseline vector, or a list of baseline vectors per
group (named list per group).
}
\examples{
library(miaTime)

data(hitchip1006)
tse <- mia::transformAssay(hitchip1006, method = "relabundance")

# Subset to speed up example
tse <- tse[, tse$subject \%in\% c("900", "934", "843", "875")]

tse2 <- addBaselineDivergence(
    tse,
    group = "subject",
    time_field = "time",
    name_divergence = "divergence_from_baseline",
    name_timedifference = "time_from_baseline",
    assay.type="relabundance",
    FUN = vegan::vegdist,
    method="bray")

tse2 <- addBaselineDivergence(
    tse,
    baseline_sample = "Sample-875",
    group = "subject",
    time_field = "time",
    name_divergence = "divergence_from_baseline",
    name_timedifference = "time_from_baseline",
    assay.type="relabundance",
    FUN = vegan::vegdist,
    method="bray")

}
