% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBaselineDivergence.R
\name{addBaselineDivergence}
\alias{addBaselineDivergence}
\alias{addBaselineDivergence,SummarizedExperiment-method}
\title{Beta diversity between the baseline and later time steps}
\usage{
addBaselineDivergence(x, ...)

\S4method{addBaselineDivergence}{SummarizedExperiment}(
  x,
  time.col,
  assay.type = "counts",
  group = NULL,
  name = "divergence",
  name.time = "time_diff",
  method = "bray",
  dimred = NULL,
  ndimred = NULL,
  dis.fun = vegan::vegdist,
  baseline_sample = NULL,
  ...
)
}
\arguments{
\item{x}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{...}{optional arguments}

\item{time.col}{\code{Character scalar}. Specifies the name of the
time series field in \code{colData}.}

\item{assay.type}{\code{Character scalar}. Specifies which assay values are
used in the dissimilarity estimation. (Default: \code{"counts"})}

\item{group}{\code{Character scalar}. Specifies the grouping
factor (name of a \code{colData} field). If given, the divergence is calculated
per group.  e.g. subject, chamber, group etc.). (Default: \code{NULL})}

\item{name}{\code{Character scalar}. Shows beta diversity between
samples. (Default: \code{"time_divergence"})}

\item{name.time}{\code{Character scalar}. Field name for adding the
time difference between samples used to calculate beta diversity.
(Default: \code{"time_difference"})}

\item{method}{\code{Character scalar}. Used to calculate the distance.
Method is passed to the function that is specified by \code{dis.fun}.
(Default: \code{"bray"})}

\item{dimred}{\code{Character scalar} or \code{integer scalar}. indicates the
reduced dimension result in \code{reducedDims} to use in the estimation.
(Default: \code{NULL})}

\item{ndimred}{\code{Integer vector}. Specifies the dimensions to use if
\code{dimred} is specified. (Default: \code{NULL})}

\item{dis.fun}{\code{Function} for dissimilarity calculation. The function must
expect the input matrix as its first argument. With rows as samples and
columns as features. (Default: \code{vegan::vegdist})}

\item{baseline_sample}{\code{Character vector}. Specifies the baseline
sample(s) to be used. If the \code{group} argument is given, this must be a
named \code{vector}; one element per group.}
}
\value{
a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
or
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
containing the sample dissimilarity and corresponding time difference between
samples (across n time steps), within each level of the grouping factor.
}
\description{
Calculates sample dissimilarity between the given baseline and other
time points, optionally within a group (subject, reaction chamber, or
similar). The corresponding time difference is returned as well.
The method operates on \code{SummarizedExperiment} objects, and the results
are stored in \code{colData}.
}
\details{
The group argument allows calculating divergence per group. Otherwise, this
is done across all samples at once.

The baseline sample/s always need to belong to the data object i.e. they
can be merged into it before
applying this function. The reason is that they need to have comparable
sample data, at least some time point
information for calculating time differences w.r.t. baseline sample.

The baseline time point is by default defined as the smallest time point
(per group). Alternatively,
the user can provide the baseline vector, or a list of baseline vectors per
group (named list per group).
}
\examples{
library(miaTime)
library(mia)

data(hitchip1006)
tse <- transformAssay(hitchip1006, method = "relabundance")

# Subset to speed up example
tse <- tse[, tse$subject \%in\% c("900", "934", "843", "875")]

tse2 <- addBaselineDivergence(
    tse,
    group = "subject",
    time.col = "time",
    name = "divergence_from_baseline",
    name.time = "time_from_baseline",
    assay.type="relabundance",
    dis.fun = vegan::vegdist,
    method="bray")

tse2 <- addBaselineDivergence(
    tse,
    baseline_sample = "Sample-875",
    group = "subject",
    time.col = "time",
    name = "divergence_from_baseline",
    name.time = "time_from_baseline",
    assay.type="relabundance",
    dis.fun = vegan::vegdist,
    method="bray")

}
