% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBaselineDivergence.R
\name{addBaselineDivergence}
\alias{addBaselineDivergence}
\alias{getBaselineDivergence}
\alias{getBaselineDivergence,SummarizedExperiment-method}
\alias{addBaselineDivergence,SummarizedExperiment-method}
\title{Beta diversity between the baseline and later time steps}
\usage{
getBaselineDivergence(x, ...)

\S4method{getBaselineDivergence}{SummarizedExperiment}(
  x,
  time.col,
  assay.type = "counts",
  reference = NULL,
  group = NULL,
  method = "bray",
  ...
)

addBaselineDivergence(x, ...)

\S4method{addBaselineDivergence}{SummarizedExperiment}(x, name = "divergence", name.time = "time_diff", ...)
}
\arguments{
\item{x}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{...}{Optional arguments passed into
\code{\link[mia:addDivergence]{mia::addDivergence()}}.}

\item{time.col}{\code{Character scalar}. Specifies a name of the column from
\code{colData} that identifies the sampling time points for the samples.}

\item{assay.type}{\code{Character scalar}. Specifies which assay values are
used in the dissimilarity estimation. (Default: \code{"counts"})}

\item{reference}{\code{Character scalar}. Specifies a name of the column from
\code{colData} that identifies the baseline samples to be used.
(Default: \code{NULL})}

\item{group}{\code{Character scalar}. Specifies a name of the column from
\code{colData} that identifies the grouping of the samples.
(Default: \code{NULL})}

\item{method}{\code{Character scalar}. Used to calculate the dissimilarity
Method is passed to the function that is specified by \code{dis.fun}.
(Default: \code{"bray"})}

\item{name}{\code{Character scalar}. Specifies a column name for storing
divergence results. (Default: \code{"divergence"})}

\item{name.time}{\code{Character scalar}. Specifies a column name for storing
time differences. (Default: \code{"time_diff"})}
}
\value{
\code{getBaselineDivergence} returns \code{DataFrame} object
containing the sample dissimilarity and corresponding time difference between
samples. \code{addBaselineDivergence}, on the other hand, returns a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object with these results in its \code{colData}.
}
\description{
Calculates sample dissimilarity between the given baseline and other
time points, optionally within a group (subject, reaction chamber, or
similar). The corresponding time difference is returned as well.
}
\details{
The group argument allows calculating divergence per group. If given, the
divergence is calculated per group.  e.g. subject, chamber, group etc.
Otherwise, this is done across all samples at once.

The baseline sample(s) always need to belong to the data object i.e. they
can be merged into it before
applying this function. The reason is that they need to have comparable
sample data, at least some time point
information for calculating time differences w.r.t. baseline sample.

The baseline time point is by default defined as the smallest time point
(per group). Alternatively,
the user can provide the baseline vector, or a list of baseline vectors per
group (named list per group).
}
\examples{
library(miaTime)
library(mia)

data(hitchip1006)
tse <- transformAssay(hitchip1006, method = "relabundance")

tse <- addBaselineDivergence(
    tse,
    group = "subject",
    time.col = "time",
    name = "divergence_from_baseline",
    name.time = "time_from_baseline",
    assay.type="relabundance",
    dis.fun = vegan::vegdist,
    method="bray")

tse <- addBaselineDivergence(
    tse,
    baseline.sample = "Sample-875",
    group = "subject",
    time.col = "time",
    name = "divergence_from_baseline",
    name.time = "time_from_baseline",
    assay.type="relabundance",
    dis.fun = vegan::vegdist,
    method="bray")

}
\seealso{
\code{\link[mia:addDivergence]{mia::addDivergence()}}
}
