% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRank.R
\name{addRank}
\alias{addRank}
\title{Rank the sample information in a \code{SummarizedExperiment} Object}
\usage{
addRank(x, field, rank_field, na.last, ties.method, ...)
}
\arguments{
\item{x}{\linkS4class{SummarizedExperiment} object}

\item{field}{A character string indicating the field aimed to be ranked}

\item{rank_field}{A single character value to name the newly ranked field}

\item{na.last}{Logical scalar: NA values are put last when \code{TRUE} ,first
when \code{FALSE}. If \code{'NA'}, they are removed; if \code{"keep"}
they are kept with rank NA.(default: \code{norm = TRUE})}

\item{ties.method}{a character string indicating the method used in ranking}

\item{...}{Allow new parameters to be defined for this function.}
}
\value{
\linkS4class{SummarizedExperiment} object with ranked field added
}
\description{
The information stored in \code{colData} in \linkS4class{SummarizedExperiment}
object is ranked by \code{addRank} function and newly ranked field is added back
to \linkS4class{SummarizedExperiment} object.
}
\examples{
library(SummarizedExperiment)
data(airway, package="airway")
se <- airway

addRank(se, field = "SampleName", rank_field = "SampleName_rank",
                                na.last ="TRUE", ties.method = "first")

}
