% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRank.R
\name{addRank}
\alias{addRank}
\title{Rank the sample information in a \code{SummarizedExperiment} Object}
\usage{
addRank(x, ...)
}
\arguments{
\item{x}{\linkS4class{SummarizedExperiment} object}

\item{...}{Allow new parameters to be defined for this function.}

\item{na.last}{NA values are put last when \code{TRUE} or put first when
\code{FALSE} or they are kept with rank NA when \code{"keep"}}

\item{ties.method}{a character string indicating the method used in ranking}
}
\value{
a list shows the ranked components of
\linkS4class{SummarizedExperiment} object
}
\description{
The information stored in \code{colData} in \linkS4class{SummarizedExperiment}
object can be ranked by \code{addRank} function.
}
\examples{
library(SummarizedExperiment)
data(airway, package="airway")
se <- airway

addRank(se, na.last ="TRUE", ties.method = "last")

}
