% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRank.R
\name{addRank}
\alias{addRank}
\title{Rank the sample information in a \code{SummarizedExperiment} Object}
\usage{
addRank(x, field, field_matrix, rank_field_name, ...)
}
\arguments{
\item{x}{\linkS4class{SummarizedExperiment} object}

\item{field}{A character value indicating the name of the matrix column that
is aimed to be ranked.}

\item{field_matrix}{A matrix that is aimed to be ranked. If the aimed matrix
string already exists in colData segment, this optional matrix doesn't need
to be added.}

\item{rank_field_name}{A character value to name the newly ranked \code{field}}

\item{...}{Allow new parameters to be defined for this function.}
}
\value{
\linkS4class{RangedSummarizedExperiment} object with ranked field
added
}
\description{
The information stored in \code{colData} in \linkS4class{SummarizedExperiment}
object is ranked by \code{addRank} function and newly ranked field is added back
to \linkS4class{SummarizedExperiment} object.
}
\examples{
library(SummarizedExperiment)
data(airway, package="airway")
se <- airway

rank <- addRank(se,field = "SampleName", rank_field_name = "SampleName_rank",
                na.last = TRUE, ties.method = "first")

#example 2: new matrix and its rank added
newmatrix <- matrix(1:8, nrow = 8, ncol = 1)

Added_field <- addRank(se, field = "NewMatrixAdded" ,
            field_matrix = newmatrix, rank_field_name = "newRankedField" ,
            na.last = TRUE , ties.method = "first")

}
