% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRank.R
\docType{methods}
\name{addRank}
\alias{addRank}
\alias{addRank-TreeSummarizedExperiment}
\alias{addRank,TreeSummarizedExperiment-method}
\title{Rank the sample information in a \code{TreeSummarizedExperiment} Object}
\usage{
addRank(x, field, rank_field_name, ...)
}
\arguments{
\item{x}{\linkS4class{TreeSummarizedExperiment} object containing time series}

\item{field}{A character value indicating the name of the matrix column that
is aimed to be ranked.}

\item{rank_field_name}{A character value to name the newly ranked \code{field}}

\item{...}{Allow new parameters to be defined for this function.}
}
\value{
\linkS4class{TreeSummarizedExperiment} object with ranked field
added
}
\description{
The information stored in \code{colData} in
\linkS4class{TreeSummarizedExperiment} object is ranked by \code{addRank}
function and newly ranked field is added back to
\linkS4class{TreeSummarizedExperiment} object.
}
\examples{
library(TreeSummarizedExperiment)
data(GlobalPatterns, package="mia")
se <- GlobalPatterns

se1 <- addRank(se,field = "SampleType", rank_field_name = "SampleType_rank",
                na.last = TRUE, ties.method = "first")

#example 2: new matrix and its rank added

se$newfield <- matrix(1:26, nrow = 26, ncol = 1)
se2 <- addRank(se, field = "newfield" ,
                rank_field_name = "newfield_rank" , na.last = TRUE ,
                ties.method = "first")

}
