% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{deprecate}
\alias{deprecate}
\alias{getTimeDivergence}
\alias{getTimeDivergence,ANY-method}
\title{These functions are deprecated. Please use other functions instead.}
\usage{
getTimeDivergence(x, ...)

\S4method{getTimeDivergence}{ANY}(x, ...)
}
\arguments{
\item{x}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{...}{Additional parameters. See dedicated function.}
}
\description{
These functions are deprecated. Please use other functions instead.
}
