% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getStability.R
\name{getStability}
\alias{getStability}
\alias{getStability,SummarizedExperiment-method}
\title{Calculate stability}
\usage{
getStability(x, ...)

\S4method{getStability}{SummarizedExperiment}(x, time.col, assay.type = "counts", group = NULL, ...)
}
\arguments{
\item{x}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{...}{additional arguments.
\itemize{
\item \code{time.interval}: \code{Integer scalar}. Indicates the increment
between time steps. By default, the function compares each sample to the
previous one. If you need to take every second, every third, or so, time
step, then increase this accordingly. (Default: \code{1L})
}}

\item{time.col}{\code{Character scalar}. Specifies a name of the column from
\code{colData} that identifies the sampling time points for the samples.}

\item{assay.type}{\code{Character scalar}. Specifies which assay values are
used in the dissimilarity estimation. (Default: \code{"counts"})}

\item{group}{\code{Character scalar}. Specifies a name of the column from
\code{colData} that identifies the grouping of the samples.
(Default: \code{NULL})}
}
\value{
Add here
}
\description{
ADD.
}
\details{
The stability is calculated by first defining reference point, \eqn{R_{f}},
for each feature. This is defined by taking mean of range of values

\deqn{R_{f} = \frac{x_{f, min} + x_{f, max}}{2}}

where \eqn{f} denotes a single feature and \eqn{x} abundance.

Then difference between consecutive time points, \eqn{\Delta_{f, x}},

\deqn{\Delta_{f, x} = ∣x_{f, t} - x_{f, t-1}∣}

and difference between the previous time point and reference,
\eqn{\Delta_{f. R}}, are calculated:

\deqn{\Delta_{f. R} = ∣x_{f, t-1} - R_{f}∣}

where \eqn{t} denotes time point.

Optionally, time difference \eqn{\Delta_{f, t}} is calculated

\deqn{\Delta_{f, t} = t_{f, t-1} - t_{f, t-1}}

The stability coefficient \eqn{s_{f}} is calculated with correlation

\deqn{s_{f} = corr(\Delta_{f, x}, \Delta_{f, R})}

or with linear model as follows

\deqn{s_{f} = \frac{ \Delta_{f, x} - \beta_{0} - \beta_{2}*\Delta_{f, t} -
\epsilon_{f} }{ \Delta_{f, R} }}
}
\examples{
library(miaTime)

# Load time series data
data(minimalgut)
tse <- minimalgut

# Apply clr transformation
tse <- transformAssay(tse, method = "rclr")

# Calculate stability single system
tse_sub <- tse[, tse[["StudyIdentifier"]] == "Bioreactor A"]
res <- getStability(tse_sub, assay.type = "rclr", time.col = "Time.hr")
res |> head()

# Calculate stability for each  system simultaneously by taking time
# difference into account
res <- getStability(
    tse, assay.type = "rclr", time.col = "Time.hr",
    group = "StudyIdentifier", mode = "lm")
res |> head()

}
\references{
Lahti L, et al. (2014) Tipping elements in the human intestinal ecosystem.
Nat Commun. doi: 10.1038/ncomms5344
}
\seealso{
\code{\link[miaTime:getBimodality]{getBimodality()}}
}
