% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStepwiseDivergence.R
\name{getStepwiseDivergence}
\alias{getStepwiseDivergence}
\alias{getTimeDivergence}
\alias{getTimeDivergence,ANY-method}
\title{Beta diversity between consecutive time steps}
\usage{
getStepwiseDivergence(
  x,
  group = NULL,
  time_field,
  time_interval = 1,
  name_divergence = "time_divergence",
  name_timedifference = "time_difference",
  assay.type = "counts",
  FUN = vegan::vegdist,
  method = "bray",
  altexp = NULL,
  dimred = NULL,
  n_dimred = NULL,
  ...
)

getTimeDivergence(x, ...)

\S4method{getTimeDivergence}{ANY}(x, ...)
}
\arguments{
\item{x}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{group}{\code{Character scalar}. Specifies the grouping
factor (name of a \code{colData} field). If given, the divergence is calculated
per group.  e.g. subject, chamber, group etc.). (Default: \code{NULL})}

\item{time_field}{\code{Character scalar}. Specifies the name of the
time series field in \code{colData}.}

\item{time_interval}{\code{Integer scalar}. Indicates the increment between time
steps. If you need to take every second, every third, or so, time step only, then
increase this accordingly. (Default: \code{1})}

\item{name_divergence}{\code{Character scalar}. Shows beta diversity between samples.
(Default: \code{"time_divergence"})}

\item{name_timedifference}{\code{Character scalar}. Field name for adding the time difference between
samples used to calculate beta diversity. (Default: \code{"time_difference"})}

\item{assay.type}{\code{Character scalar}. Specifies which assay values are used in
the dissimilarity estimation. (Default: \code{"counts"})}

\item{FUN}{\code{Function} for dissimilarity calculation. The function must
expect the input matrix as its first argument. With rows as samples
and columns as features. (Default: \code{vegan::vegdist})}

\item{method}{\code{Character scalar}. Used to calculate the distance. Method is
passed to the function that is specified by \code{FUN}. (Default: \code{"bray"})}

\item{altexp}{\code{Character scalar} or \code{integer scalar}. Specifies the alternative experiment
containing the input data. (Default: \code{NULL})}

\item{dimred}{\code{Character scalar} or \code{integer scalar}. indicates the reduced dimension
result in \code{reducedDims} to use in the estimation. (Default: \code{NULL})}

\item{n_dimred}{\code{Integer vector}. Specifies the dimensions to use if
\code{dimred} is specified. (Default: \code{NULL})}

\item{...}{Arguments to be passed}
}
\value{
a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
or
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
containing the sample dissimilarity and corresponding time difference between
samples (across n time steps), within each level of the grouping factor.
}
\description{
Calculates sample dissimilarity between consecutive time points (t, t+i),
within a group (subject, reaction chamber, or similar). The corresponding
time difference is returned as well. The method operates on
\code{SummarizedExperiment} objects, and the results are stored in \code{colData}.
}
\examples{
#library(miaTime)
library(TreeSummarizedExperiment)

data(hitchip1006)
tse <- mia::transformCounts(hitchip1006, method = "relabundance")

# Subset to speed up example
tse <- tse[, colData(tse)$subject \%in\% c("900", "934", "843", "875")]

# Using vegdist for divergence calculation, one can pass
# the dissimilarity method from the vegan::vegdist options
# via the "method" argument
tse2 <- getStepwiseDivergence(tse, group = "subject",
                              time_interval = 1,
                              time_field = "time",
                              assay.type="relabundance",
                              FUN = vegan::vegdist,
                              method="bray")

}
