% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimeDivergence.R
\name{getTimeDivergence}
\alias{getTimeDivergence}
\title{Beta diversity calculation between samples at the interval of n time steps}
\usage{
getTimeDivergence(
  se,
  sample_field,
  sample_id,
  time_field,
  time_interval,
  transposed = FALSE
)
}
\arguments{
\item{se}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object or
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}}

\item{sample_field}{column vector containing the all individuals in
\code{colData} field}

\item{sample_id}{character indicating a individual inside of \code{sample_field}}

\item{time_field}{column vector containing the time information
in \code{colData} field}

\item{time_interval}{integer value indicating the increment between the time
steps. It can be 1 or higher than 1.}

\item{transposed}{logical scalar assigning samples to rows if they are in
columns (default: \code{transposed = FALSE})}
}
\value{
a list containing the beta diversity in matrix and a numeric vector
containing the time difference between the samples
}
\description{
The dissimilarity (beta distance) is calculated in steps of n for each
individual. "n" can be 1 or greater than 1. The time difference between
the time points used for beta diversity calculation is also given as output.
}
\examples{
library(microbiomeDataSets)
tse <- GrieneisenTSData()

BaboonDivergence <- getTimeDivergence(tse, sample_field = "baboon_id",
                                    sample_id = "Baboon_1" ,
                                    time_field = "collection_date",
                                    time_interval =2,
                                    transposed = FALSE )

}
