% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimeDivergence.R
\name{getTimeDivergence}
\alias{getTimeDivergence}
\title{Beta diversity calculation within individuals over time}
\usage{
getTimeDivergence(
  se,
  group,
  time_field,
  time_interval = 1,
  name_divergence = "time_divergence",
  name_timedifference = "time_difference",
  abund_values = "counts",
  distfun = vegan::vegdist
)
}
\arguments{
\item{se}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{group}{a single character value for specifying which grouping
factor is used (name of a \code{colData} field).}

\item{time_field}{a single character value, specifying the name of the
time series field in \code{colData}.}

\item{time_interval}{integer value indicating the increment between time
steps (default: 1)}

\item{name_divergence}{a column vector showing beta diversity between samples
over n time intervals (default: \code{name_divergence = "time_divergence"})}

\item{name_timedifference}{field name for adding the time difference between
samples used to calculate beta diversity
(default: \code{name_timedifference = "time_difference"})}

\item{abund_values}{character indicating which assay values are used in
the dissimilarity estimation (default: \code{abund_values = "counts"})}

\item{distfun}{a function to calculate beta diversity.
(default: \code{distfun = vegan::vegdist})}
}
\value{
a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
or
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
containing the sample dissimilarity and corresponding time difference between
samples (across n time steps), within each level of the grouping factor.
}
\description{
This method calculates sample dissimilarity between consecutive time steps
(step size n=1 by default), within a group (subject, reaction chamber,
or similar).
The corresponding time difference is returned as well. The method operates on
\code{SummarizedExperiment} objects, and the results are stored in \code{colData}.
}
\examples{
#library(miaTime)
data(hitchip1006)
se <- hitchip1006

# Subset to speed up example
se <- se[,878:1011]
se2 <- getTimeDivergence(se, group = "subject",
                             time_interval = 1,
                             time_field = "time")

}
