% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimeDivergence.R
\name{getTimeDivergence}
\alias{getTimeDivergence}
\title{Time difference between Nth time steps}
\usage{
getTimeDivergence(se, time_field, starting_position = 1, time_interval)
}
\arguments{
\item{se}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{time_field}{column vector containing the time information
in \code{colData} field inside of \linkS4class{SummarizedExperiment} object}

\item{starting_position}{integer value that defines from which time step the
operation begins (default: \code{starting_position = 1})}

\item{time_interval}{integer value indicating the increment between the time
steps. It can be 1 or higher than 1.}
}
\value{
\code{colData} field in \linkS4class{SummarizedExperiment} object
}
\description{
The time difference can be calculated between successive time steps, with
steps n >= 1.
}
\examples{
library(miaTime)
data(hitchip1006)
se <- hitchip1006

seventeenTime <- getTimeDivergence(se, time_field = "time" , time_interval = 17)


}
