% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimeDivergence.R
\name{getTimeDivergence}
\alias{getTimeDivergence}
\title{Beta diversity calculation within individuals in the interval of n time steps}
\usage{
getTimeDivergence(se, field, time_field, time_interval)
}
\arguments{
\item{se}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object or
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}}

\item{field}{a single character value for specifying which grouping
factor chosen from columns of \code{colData} field.}

\item{time_field}{a single character value for specifying time series
vector given in \code{colData} field.}

\item{time_interval}{integer value indicating the increment between the time
steps. It can be 1 or higher than 1. (default: \code{time_interval = 1})}
}
\value{
a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
or
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
containing the beta diversity and time difference between samples
in n time steps
}
\description{
The dissimilarity (beta diversity) is calculated in time steps of n, between
the samples of subject(individuals).Time interval can be 1 or greater than 1.
The time difference between n points is also calculated. Given input,
\code{SummarizedExperiment} or \code{TreeSummarizedExperiment} object, returns with
the calculations inserted in columns of \code{colData} field.
}
\examples{
library(miaTime)
data(hitchip1006)

hitchipTime <- getTimeDivergence(hitchip1006, field = "subject",
                                   time_interval = 3,
                                   time_field = "time")

}
