% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miaTime.R
\docType{package}
\name{miaTime-package}
\alias{miaTime-package}
\alias{miaTime}
\title{\code{miaTime} Package.}
\description{
\code{miaTime} implements time series methods in \link[mia]{mia} ecosystem.
}
\seealso{
\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment} class
}
