% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miaTime.R
\docType{package}
\name{miaTime-package}
\alias{miaTime-package}
\alias{miaTime}
\title{\code{miaTime} Package.}
\description{
\code{miaTime} implements time series methods in \link[mia]{mia} ecosystem.
}
\seealso{
\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment} class
}
\author{
\strong{Maintainer}: Yagmur Simsek \email{yagmur.simsek@hsrw.org}

Authors:
\itemize{
  \item Leo Lahti \email{leo.lahti@iki.fi} (\href{https://orcid.org/0000-0001-5537-637X}{ORCID})
}

Other contributors:
\itemize{
  \item Sudarshan Shetty \email{sudarshanshetty9@gmail.com} [contributor]
  \item Chouaib Benchraka [contributor]
  \item Tuomas Borman \email{tuomas.v.borman@utu.fi} (\href{https://orcid.org/0000-0002-8563-8884}{ORCID}) [contributor]
  \item Muluh Muluh [contributor]
}

}
