% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeDifferences.R
\name{timeDifferences}
\alias{timeDifferences}
\title{Consecutive time difference of numeric values can be calculated
as well as the time difference based on the certain time point with
parameter \code{baseline}.}
\usage{
timeDifferences(time_field, baseline = 1)
}
\arguments{
\item{time_field}{: vector containing time points}

\item{baseline}{: integer, value to calculate the distance from certain time
point. (default: \code{baseline = 1})}
}
\value{
a list object containing consecutive time differences and
the distance from selected time point to each time point.
}
\description{
Consecutive time difference of numeric values can be calculated
as well as the time difference based on the certain time point with
parameter \code{baseline}.
}
\examples{
data(hitchip1006)

diffHitchip <- timeDifferences(time_field = colData(hitchip1006)[,"time"],
                                             baseline = 1)

##Each of these fields in the list object can be applied to the
#\linkS4class{SummarizedExperiment} object as new fields.

colData(hitchip1006)$time_diff <- diffHitchip$Consecutive
colData(hitchip1006)$base_diff <- diffHitchip$Baseline

}
