% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeToDate.R
\name{timeToDate}
\alias{timeToDate}
\title{Numerically provided time points are converted to a character object in the
format "\%H:\%M:\%S".}
\usage{
timeToDate(data, time_field, unit = "mins")
}
\arguments{
\item{data}{: dataset with numeric time points in it}

\item{time_field}{: character, time points field in the dataset}

\item{unit}{: character, indicating the unit of given times
(default: \code{unit = "mins"})}
}
\value{
a list object containing time in "\%H:\%M:\%S" format.
}
\description{
Numerically provided time points are converted to a character object in the
format "\%H:\%M:\%S".
}
\examples{
data(hitchip1006)
hitchipDate <- timeToDate(hitchip1006, time_field = "time", unit = "mins")

##This can be added back to \linkS4class{SummarizedExperiment} object that
#is used to store `hitchip1006` dataset.

colData(hitchip1006)$clock <- hitchipDate

}
